Adding master libraries and code chunks
Trying to centralize macros and code chunks that keep turning up in mods and/or may be useful outside of mine. This is phase one
where I'm trying to consolidate these from my own mods; next I'll comb through the semi-useful macro thread on G3 and add those
here as well.

Note that some of these will say they're used in a mod at (some file path) and may not actually be there yet. I've made some
changes locally and pushed these to Github, as I don't want to just change a bunch of stuff for the sake of change, so you'll
see these in the next official releases. For now I'm simply symlinked files (locally) from other repos to the files here, and
these will eventually propagate.

Note also that these are copy-pasted, so it's entirely possible I've left a */ somewhere that may throw a syntax error. That's
next on the to-do list so (for now at least) don't just blindly use these. If you do hit a bug, let me know or send a pull
request.

If you're reading this, you're probably a modder and have some of these yourself. Feel free to add any libraries or mutual code
chunks you think may be useful to a broader audience via pull request.